Adds NA handling to thermo.depth function.
This change allows temperature profiles with NAs to be used as inputs. It does this by removing the depth with no measured temperature from the depth vector as well as the NA from the temperature profile. It will also generate a warning, which can be turned off by setting warn.on.NA = FALSE.